+++ tests/Agent/IntegrationTests/SharedApplications/Common/MultiFunctionApplicationHelpers/NetStandardLibraries/CosmosDB/StoredProcedures/HelloWorldStoredProc.js
@@ -1,0 +1,6 @@
+﻿function hello () {
+    var context = getContext();
+    var response = context.getResponse();
+    response.setBody("Hello, World");
+}
+