+++ tests/Agent/UnitTests/Core.UnitTest/Spans/SpanStreamingServiceTests.cs
@@ -716,7 +716,7 @@ namespace NewRelic.Agent.Core.Spans.Tests
         }
 
         [Test]
-        [Ignore("This test is flickering in our CI", Until = "2021-11-08 00:00:00Z")]
+        [Ignore("This test is flickering in our CI", Until = "2022-06-01 00:00:00Z")]
         public void ShuttingDownTheDataStreamingService_ShouldShutdownResponseStream()
         {
             var signalIsDone = new ManualResetEventSlim();
@@ -964,7 +964,7 @@ namespace NewRelic.Agent.Core.Spans.Tests
 
 
         [Test]
-        [Ignore("This test is flickering in our CI", Until = "2021-11-08 00:00:00Z")]
+        [Ignore("This test is flickering in our CI", Until = "2022-06-01 00:00:00Z")]
         public void SupportabilityMetrics_ItemsSent_BatchSizeAndCount()
         {
             const int maxBatchSize = 17;